World providers will no longer crash if no worlds are found Added handling for world/location providers returning null Added more debug information for UUID and Name world providers Fixed location proiders causing integer overflow during default Y value calculations

+++ src/com/thizthizzydizzy/resourcespawner/provider/location/CircleLocationProvider.java
@@ -29,9 +29,9 @@ public class CircleLocationProvider implements LocationProvider{
         if(ResourceSpawnerCore.debug)System.out.println("z: "+originZ);
         radius = json.getInt("radius", 0);
         if(ResourceSpawnerCore.debug)System.out.println("radius: "+radius);
-        minY = json.getInt("min_y", Integer.MIN_VALUE);
+        minY = json.getInt("min_y", Short.MIN_VALUE);
         if(ResourceSpawnerCore.debug)System.out.println("min y: "+minY);
-        maxY = json.getInt("max_y", Integer.MAX_VALUE);
+        maxY = json.getInt("max_y", Short.MAX_VALUE);
         if(ResourceSpawnerCore.debug)System.out.println("max y: "+maxY);
         if(maxY<minY)throw new IllegalArgumentException("max_y must be greater than or equal to min_y!");
         verticalDistribution = plugin.getDistribution(json.getString("vertical_distribution", "even"));
